Actually, create a dedicated graphics/selection.rs
